Inventory Database and Website
This is a website I have created to track Trading Cards.

This website uses Express Router to direct Controller functions, achieving full CRUD access to a PostgreSQL database with 4 tables. The Controllers render pages accordingly using EJS.

I have included a script to load up 6 examples of cards and the table fields.

I have included a .env.example outlining the Environment Variables needed.

This page contains minimal styling as it's primary purpose is to act as a proof of knowledge for The Odin Project's Node course so far. I may revisit and style at a later stage.